requirements.txt
streamlit
openai
chromadb
langchain
tiktoken
pdfplumber
python-dotenv
pyttsx3
spacy
mistralai